Move database utility classes to a separate module
Move the `InsertDump`, `ConnectionWrapper`, and `DatabaseContext`
classes into a new `db.utils` module.
